added job_skilltag function and added dashboard files, cleaned up some code
