+++ controllers/index.js
@@ -1,12 +1,12 @@
-const pg = require('pg');
-const express = require('express')
-const app = express();
-const ejs = require('ejs')
-const bodyParser = require('body-parser');
+// const pg = require('pg');
+// const express = require('express')
+// const app = express();
+// const ejs = require('ejs')
+// const bodyParser = require('body-parser');
 const { pool } = require('../models/database');
 
-app.set('view engine', 'ejs')
-app.use(express.urlencoded({ extended: true }))
+// app.set('view engine', 'ejs')
+// app.use(express.urlencoded({ extended: true }))
 
 //routes
 exports.getIndex = (req, res, next) => {
