+++ models/knex.js
@@ -1,8 +1,5 @@
-//database connection file creation
-const environment = process.env.NODE_ENV  || 'development';
-const config = require('./knexfile');
-const environmentConfig = config[enviornment];
 const knex = require('knex');
-const connection = knex(environmentConfig);
-
-module.exports = connection;
+const knexfile = require('../knexfile');
+const env = process.env.NODE_ENV || 'development';
+const configOptions = knexfile[env];
+module.exports = knex(configOptions);
