+++ routes/index.js
@@ -1,7 +1,7 @@
 //==== Routers set up ====//
 
 //require modules
-const path = require('path');
+// const path = require('path');
 const express = require('express');
 const router = express.Router();
